requirements.txt
certifi
chardet
click
colorama
Flask
gunicorn
idna
itsdangerous
Jinja2
MarkupSafe
python-dotenv
requests
urllib3
Werkzeug
